+++ corehq/apps/accounting/tests/test_invoicing.py
@@ -32,6 +32,7 @@ from corehq.apps.smsbillables.models import (
     SmsUsageFee,
     SmsUsageFeeCriteria,
 )
+from corehq.apps.smsbillables.tests.generator import arbitrary_sms_billables_for_domain
 
 
 class BaseInvoiceTestCase(BaseAccountingTest):
@@ -495,10 +496,10 @@ class TestSmsLineItem(BaseInvoiceTestCase):
         - total and subtotals are 0.0
         """
         num_sms = random.randint(0, self.sms_rate.monthly_limit/2)
-        generator.arbitrary_sms_billables_for_domain(
+        arbitrary_sms_billables_for_domain(
             self.subscription.subscriber.domain, self.sms_date, num_sms, direction=INCOMING
         )
-        generator.arbitrary_sms_billables_for_domain(
+        arbitrary_sms_billables_for_domain(
             self.subscription.subscriber.domain, self.sms_date, num_sms, direction=OUTGOING
         )
         sms_line_item = self._create_sms_line_item()
@@ -524,7 +525,7 @@ class TestSmsLineItem(BaseInvoiceTestCase):
         - total and subtotals are greater than zero
         """
         num_sms = random.randint(self.sms_rate.monthly_limit + 1, self.sms_rate.monthly_limit + 2)
-        billables = generator.arbitrary_sms_billables_for_domain(
+        billables = arbitrary_sms_billables_for_domain(
             self.subscription.subscriber.domain, self.sms_date, num_sms
         )
         sms_line_item = self._create_sms_line_item()
@@ -572,12 +573,12 @@ class TestSmsLineItem(BaseInvoiceTestCase):
         for billable in SmsBillable.objects.all():
             _set_billable_date_sent_day(billable, 1)
 
-        half_charged_billable = generator.arbitrary_sms_billables_for_domain(
+        half_charged_billable = arbitrary_sms_billables_for_domain(
             self.subscription.subscriber.domain, self.sms_date, 1, multipart_count=2
         )[0]
         _set_billable_date_sent_day(half_charged_billable, 2)
 
-        fully_charged_billable = generator.arbitrary_sms_billables_for_domain(
+        fully_charged_billable = arbitrary_sms_billables_for_domain(
             self.subscription.subscriber.domain, self.sms_date, 1, multipart_count=random.randint(2, 5)
         )[0]
         _set_billable_date_sent_day(fully_charged_billable, 3)
@@ -606,7 +607,7 @@ class TestSmsLineItem(BaseInvoiceTestCase):
         while True:
             multipart_count = random.randint(1, 5)
             if count_parts + multipart_count <= total_parts:
-                generator.arbitrary_sms_billables_for_domain(
+                arbitrary_sms_billables_for_domain(
                     self.subscription.subscriber.domain, self.sms_date, 1, multipart_count=multipart_count
                 )
                 count_parts += multipart_count
@@ -614,7 +615,7 @@ class TestSmsLineItem(BaseInvoiceTestCase):
                 break
         remaining_parts = total_parts - count_parts
         if remaining_parts > 0:
-            generator.arbitrary_sms_billables_for_domain(
+            arbitrary_sms_billables_for_domain(
                 self.subscription.subscriber.domain, self.sms_date, 1, multipart_count=remaining_parts
             )
